KEL-75 H2Interface (which implements HousingDatabaseInterface) now uses 3 services: HouseService, OccupantService, and AssignmentService. Detailed implementation of the methods from H2Interface is moved to those service classes.
