+++ src/main/java/com/example/housingmanagement/api/OccupantRepositoryJPA.java
@@ -6,5 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface OccupantRepositoryJPA extends JpaRepository<OccupantInternalEntity, Integer> {
     OccupantInternalEntity findByFirstNameAndLastName(String firstName, String lastName);
 
+    void deleteByFirstNameAndLastName(String firstName, String lastName);
+
     boolean existsByFirstNameAndLastName(String firstName, String lastName);
 }
